requirements.txt
moviepy
openai-whisper
ffmpeg-python
rich
click
google-generativeai
python-dotenv
